rebase rename
-  changing all the files/folders to lowercase with suffix
