+++ src/main.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./App.jsx";
+import App from "./app-1.jsx";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import "./index.css";
